Remove the dependency from `protoc-plugin` to `model-compiler` by moving the API files into a shared `protoc-api` module. Restore using the `protoc-plugin` JAR in the `model-compiler` testing.
