+++ tools/model-compiler/src/main/java/io/spine/tools/gradle/compiler/Extension.java
@@ -26,8 +26,8 @@ import groovy.lang.Closure;
 import io.spine.code.fs.java.DefaultJavaProject;
 import io.spine.code.gen.Indent;
 import io.spine.tools.gradle.GradleExtension;
-import io.spine.tools.gradle.compiler.protoc.GeneratedInterfaces;
-import io.spine.tools.gradle.compiler.protoc.GeneratedMethods;
+import io.spine.tools.protoc.GeneratedInterfaces;
+import io.spine.tools.protoc.GeneratedMethods;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
