+++ tools/protoc-api/src/main/java/io/spine/tools/protoc/GeneratedInterfaces.java
@@ -18,13 +18,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle.compiler.protoc;
+package io.spine.tools.protoc;
 
 import io.spine.annotation.Internal;
 import io.spine.base.EventMessage;
 import io.spine.code.java.ClassName;
-import io.spine.tools.protoc.AddInterfaces;
-import io.spine.tools.protoc.UuidConfig;
 import org.checkerframework.checker.signature.qual.FullyQualifiedName;
 
 import static com.google.common.base.Preconditions.checkNotNull;
