+++ tools/protoc-api/src/test/java/io/spine/tools/protoc/MessageSelectorFactoryTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle.compiler.protoc;
+package io.spine.tools.protoc;
 
 import com.google.common.collect.ImmutableMap;
 import io.spine.code.proto.FileName;
