+++ tools/protoc-plugin/src/test/java/io/spine/tools/protoc/PluginTest.java
@@ -28,10 +28,6 @@ import io.spine.code.fs.java.SourceFile;
 import io.spine.code.java.ClassName;
 import io.spine.code.proto.OptionExtensionRegistry;
 import io.spine.protobuf.ValidatingBuilder;
-import io.spine.tools.gradle.compiler.protoc.GeneratedInterfaces;
-import io.spine.tools.gradle.compiler.protoc.GeneratedMethods;
-import io.spine.tools.gradle.compiler.protoc.MessageSelectorFactory;
-import io.spine.tools.gradle.compiler.protoc.PatternSelector;
 import io.spine.tools.protoc.given.TestInterface;
 import io.spine.tools.protoc.given.TestMethodFactory;
 import io.spine.tools.protoc.given.UuidMethodFactory;
@@ -54,9 +50,9 @@ import java.util.List;
 
 import static com.google.common.collect.Iterables.getOnlyElement;
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.tools.gradle.compiler.protoc.MessageSelectorFactory.prefix;
-import static io.spine.tools.gradle.compiler.protoc.MessageSelectorFactory.regex;
-import static io.spine.tools.gradle.compiler.protoc.MessageSelectorFactory.suffix;
+import static io.spine.tools.protoc.MessageSelectorFactory.prefix;
+import static io.spine.tools.protoc.MessageSelectorFactory.regex;
+import static io.spine.tools.protoc.MessageSelectorFactory.suffix;
 import static io.spine.tools.protoc.given.CodeGeneratorRequestGiven.protocConfig;
 import static io.spine.tools.protoc.given.CodeGeneratorRequestGiven.requestBuilder;
 import static java.util.stream.Collectors.toList;
