+++ tools/protoc-plugin/src/test/java/io/spine/tools/protoc/SpineProtoGeneratorTest.java
@@ -27,9 +27,6 @@ import com.google.protobuf.compiler.PluginProtos.CodeGeneratorRequest;
 import com.google.protobuf.compiler.PluginProtos.CodeGeneratorResponse;
 import com.google.protobuf.compiler.PluginProtos.CodeGeneratorResponse.File;
 import io.spine.code.java.ClassName;
-import io.spine.tools.gradle.compiler.protoc.GeneratedInterfaces;
-import io.spine.tools.gradle.compiler.protoc.GeneratedMethods;
-import io.spine.tools.gradle.compiler.protoc.MessageSelectorFactory;
 import io.spine.tools.protoc.given.TestInterface;
 import io.spine.tools.protoc.given.UuidMethodFactory;
 import io.spine.type.MessageType;
